refactor(*) remove unused components
