+++ lib/components/Datasets.js
@@ -17,10 +17,6 @@ import Spinner from './Spinner'
 import Header from './Header'
 import Button from './Button'
 import { addActiveToLink } from '../utils/links.js'
-import AddDatasetContainer from '../containers/AddDataset'
-// import Search from './Search'
-
-const ADD_DATASET_MODAL = 'ADD_DATASET_MODAL'
 
 class Datasets extends Base {
   constructor (props) {
@@ -32,19 +28,10 @@ class Datasets extends Base {
       error: false
     };
 
-    // TODO - restore search
-    // this.debounceRunDatasetSearch = debounce((searchString) => {
-    //   this.props.runDatasetSearch(searchString)
-    // }
-    // , 250);
-
     [
-      // 'handleDatasetSearch',
       'handleLoadDatasetsError',
       'handleReloadDatasets',
-      'handleAddItem',
       'handleLoadNextPage',
-      'handleAddDataset',
       'renderLocalDatasets'
     ].forEach((m) => { this[m] = this[m].bind(this) })
   }
@@ -80,34 +67,6 @@ class Datasets extends Base {
     this.props.loadDatasets(this.props.sessionProfileId, this.props.nextPage - 1).then(action => this.handleLoadDatasetsError(action))
   }
 
-  handleAddItem () {
-    this.props.showModal(ADD_DATASET_MODAL, this, {}, true)
-  }
-
-  // TODO - restore search
-  // handleDatasetSearch (searchString) {
-  //   this.setState({ loading: true })
-  //   this.props.setDatasetSearch(searchString)
-  //   this.debounceRunDatasetSearch(searchString)
-  // }
-
-  modal (name, data = {}) {
-    switch (name) {
-      case ADD_DATASET_MODAL:
-        return <AddDatasetContainer />
-      default:
-        return undefined
-    }
-  }
-
-  handleAddDataset (peername, name, path, peer) {
-    return peer ? () => {
-      this.props.addDataset(peername, name, path).then(() => {
-        this.props.loadDatasets(this.props.sessionProfileId)
-      })
-    } : undefined
-  }
-
   renderLocalDatasets (css) {
     const { message, loading, error } = this.state
     const { datasets, palette, smallItems } = this.props
@@ -124,7 +83,6 @@ class Datasets extends Base {
               loading={this.props.loading}
               fetchedAll={this.props.fetchedAll}
               onClick={this.handleLoadNextPage}
-              onAdd={this.handleAddDataset}
               type='datasets'
             />
         }
