+++ lib/components/Hash.js
@@ -28,10 +28,10 @@ export default class Hash extends Base {
     const palette = defaultPalette
     return {
       prefix: {
-        color: palette.neutralMuted
+        color: palette.neutral
       },
       hash: {
-        color: palette.neutralBold
+        color: palette.neutral
       },
       old: {
         color: palette.d,
