+++ lib/components/StatsBar.js
@@ -1,27 +1,44 @@
 import React from 'react'
 import PropTypes from 'prop-types'
-import Hash from './Hash'
 
 import { defaultPalette } from '../propTypes/palette'
-import { relDate } from '../utils/date'
 
 import Base from './Base'
 
-export default class CommitLine extends Base {
+export default class StatsBar extends Base {
+  constructor (props) {
+    super(props);
+    [
+      'renderStat',
+      'parseDate'
+    ].forEach((m) => { this[m] = this[m].bind(this) })
+  }
+
+  parseDate (datetime) {
+    if (!datetime) {
+      return 'no date given'
+    }
+    const date = new Date(datetime)
+    const options = { month: 'short', day: 'numeric', year: 'numeric' }
+    return date.toLocaleDateString('en-US', options)
+  }
+
+  renderStat (stat, space, css, index, style) {
+    if (stat && stat.name) {
+      const num = stat.value > 1000 ? Math.round(stat.value / 1000) + 'K' : stat.value
+      return <span key={index} className={css(space)}><strong>{num}</strong><br />{stat.name}</span>
+    }
+  }
+
   template (css) {
-    const { muted, extraSpace, large, style, path, commit } = this.props
+    const { stats, muted, extraSpace, large, style, updated } = this.props
     const space = extraSpace ? 'extraSpace' : 'space'
     const displayLarge = large ? '' : 'stats'
     const displayMuted = muted ? 'muted' : ''
     return (
       <div className={`${css('wrap')} stats-line-large ${displayLarge} ${displayMuted ? css(displayMuted) : undefined}`} style={style} >
-        <div className={css('left')}>
-          <p>{commit.title}</p>
-        </div>
-        <div className={css('right')}>
-          <Hash hash={path} />
-          {commit.created && <span className={css(space)}>{relDate(commit.created)}</span> }
-        </div>
+        {stats.map((stat, index) => this.renderStat(stat, space, css, index))}
+        {updated ? <span className={css(space)}><strong>{this.parseDate(updated)}</strong><br />updated</span> : undefined}
       </div>
     )
   }
@@ -32,18 +49,16 @@ export default class CommitLine extends Base {
       wrap: {
         display: 'flex',
         width: '100%',
-        marginTop: 10,
+        marginTop: 40,
         paddingTop: 12,
         paddingBottom: 10,
+        border: '1px solid #eee',
         borderRadius: 3
       },
-      left: {
-        flex: '1 1 80px'
-      },
-      right: {
+      space: {
         marginRight: 15,
         flex: '2 1 80px',
-        textAlign: 'right',
+        textAlign: 'center',
         color: palette.neutralBold
       },
       extraSpace: {
@@ -59,14 +74,14 @@ export default class CommitLine extends Base {
   }
 }
 
-CommitLine.propTypes = {
+StatsBar.propTypes = {
   stats: PropTypes.arrayOf(PropTypes.object).isRequired,
   muted: PropTypes.bool,
   extraSpace: PropTypes.bool,
   style: PropTypes.object
 }
 
-CommitLine.defaultProps = {
+StatsBar.defaultProps = {
   style: {},
   stats: [],
   muted: false,
