+++ lib/components/TabPanel.js
@@ -3,21 +3,18 @@ import PropTypes from 'prop-types'
 import Base from './Base'
 
 import { defaultPalette } from '../propTypes/palette'
-import Search from './Search'
 
 export default class TabPanel extends Base {
   constructor (props) {
     super(props)
 
     this.state = {
-      index: 0,
-      searchString: ''
+      index: 0
     };
 
     [
       'handleSetPanel',
-      'setLocationBarText',
-      'renderSearch'
+      'setLocationBarText'
     ].forEach((m) => { this[m] = this[m].bind(this) })
   }
 
@@ -29,20 +26,8 @@ export default class TabPanel extends Base {
     }
   }
 
-  setLocationBarText (searchString) {
-    this.setState({ searchString })
-  }
-
-  renderSearch (css) {
-    return (
-      <div className={css('search')}>
-        <Search searchString={this.state.searchString} onChange={this.setLocationBarText} placeholder='search data' />
-      </div>
-    )
-  }
-
   template (css) {
-    const { index, expanded, labels = [], components, onSelectPanel, onToggleExpand, search, clearBackground } = this.props
+    const { index, expanded, labels = [], components, onSelectPanel, onToggleExpand, clearBackground } = this.props
     const panel = (index !== undefined) ? index : this.state.index
     const component = components[panel]
     const cssWrap = clearBackground ? css('wrap') : css('wrap', 'background')
@@ -60,8 +45,6 @@ export default class TabPanel extends Base {
           }
           )}
         </div>
-
-        {search ? this.renderSearch(css) : undefined}
         <div className={`${css('content')} ${clearBackground ? css('contentSink') : ''}`}>
           {component}
         </div>
