+++ lib/components/dataset/DatasetHeader.js
@@ -6,6 +6,7 @@ import DatasetRefProps from '../../propTypes/datasetRefProps'
 import Button from '../Button'
 import DatasetName from '../DatasetName'
 import Header from '../Header'
+import Hash from '../Hash'
 
 import { defaultPalette } from '../../propTypes/palette'
 import { addActiveToLink } from '../../utils/links.js'
@@ -29,11 +30,11 @@ class DatasetHeader extends Base {
   }
 
   renderRef (css) {
-    const { peername, name } = this.props
+    const { peername, name, datasetRef, isLatestDataset } = this.props
     return (
       <div>
         <DatasetName peername={peername} name={name || 'unnamed dataset'} large link style={{ display: 'inline-block', marginTop: 15 }} />
-        {/* <Hash hash={datasetRef.path} style={{ marginBottom: 15 }} old={!isLatestDataset} /> */}
+        <Hash hash={datasetRef.path} style={{ marginBottom: 15 }} old={!isLatestDataset} />
       </div>
     )
   }
