+++ lib/components/dataset/Overview.js
@@ -4,7 +4,7 @@ import PropTypes from 'prop-types'
 import MetaProps from '../../propTypes/metaProps.js'
 import Base from '../Base'
 import TagList from '../TagList'
-import StatsLineLarge from '../StatsLineLarge'
+import StatsBar from '../StatsBar'
 import CommitItem from '../item/CommitItem'
 import Button from '../Button'
 
@@ -175,7 +175,7 @@ export default class Overview extends Base {
         {/* md && this.renderTitle(css, 'Meta', 'meta') */}
         {this.renderMeta(css, md, name)}
         {/* structure && updated && this.renderTitle(css, 'Structure', 'structure') */}
-        {structure && updated && <StatsLineLarge stats={this.stats(structure)} updated={updated} />}
+        {structure && updated && <StatsBar stats={this.stats(structure)} updated={updated} />}
         {/* commit && this.renderTitle(css, 'Commit', 'commit') */}
         {commit && <CommitItem data={datasetRef} profile={profile} /> }
         {/* this.renderFields(md, css) */}
