+++ lib/components/profile/NoProfile.js
@@ -1,8 +1,7 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 
-import PageHeader from '../PageHeader'
-import ReadOnlyHeader from '../ReadOnlyHeader'
+import Header from '../Header'
 
 import { defaultPalette } from '../../propTypes/palette'
 import Base from '../Base'
@@ -16,24 +15,11 @@ export default class NoProfile extends Base {
     ].forEach((m) => { this[m] = this[m].bind(this) })
   }
 
-  renderHeader () {
-    const { onGoBack, sessionProfileId } = this.props
-    if (sessionProfileId) {
-      return <PageHeader
-        onGoBack={onGoBack}
-      />
-    } else {
-      return <ReadOnlyHeader
-        onGoBack={onGoBack}
-      />
-    }
-  }
-
   template (css) {
     const session = this.props.sessionProfileId ? 'session' : 'noSession'
     return (
       <div>
-        { this.renderHeader() }
+        <Header text='No Profile' />
         <div className={css('page', session)}>
           <div className={css('center')}>
             <h3>Profile '{this.props.peername}' not found or cannot be loaded</h3>
@@ -44,10 +30,9 @@ export default class NoProfile extends Base {
   }
 
   styles () {
-    const palette = defaultPalette
     return {
       page: {
-        background: palette.background,
+        background: defaultPalette.background,
         position: 'absolute',
         width: '100%',
         minHeight: 1000,
