+++ lib/components/profile/Profile.js
@@ -9,7 +9,6 @@ import TabPanel from '../TabPanel'
 import List from '../List'
 import DatasetItem from '../item/DatasetItem'
 import Spinner from '../Spinner'
-import NameDataset from '../NameDataset'
 import NoProfile from './NoProfile'
 import ProfileSidebar from './ProfileSidebar'
 import ProfileEditorContainer from '../../containers/ProfileEditor'
@@ -17,7 +16,6 @@ import ReadOnly from '../ReadOnly'
 
 import Base from '../Base'
 
-const ADD_PROFILE_DATASET_MODAL = 'ADD_PROFILE_DATASET_MODAL'
 const EDIT_PROFILE_MODAL = 'EDIT_PROFILE_MODAL'
 
 export default class Profile extends Base {
@@ -103,15 +101,6 @@ export default class Profile extends Base {
     this.props.setProfilePhoto(files)
   }
 
-  handleShowAddProfileDatasetModal (datasetRef) {
-    const data = {
-      datasetRef,
-      handleNameDataset: this.handleAddDataset,
-      hideModal: this.props.hideModal
-    }
-    this.props.showModal(ADD_PROFILE_DATASET_MODAL, this, data, false)
-  }
-
   handleShowEditProfileModal () {
     this.props.showModal(EDIT_PROFILE_MODAL, this, this.props.profile, true, false)
   }
@@ -125,8 +114,6 @@ export default class Profile extends Base {
 
   modal (name, data = {}) {
     switch (name) {
-      case ADD_PROFILE_DATASET_MODAL:
-        return <NameDataset data={data} />
       case EDIT_PROFILE_MODAL:
         return <ProfileEditorContainer />
       default:
