+++ lib/components/profile/ProfileSidebar.js
@@ -14,7 +14,6 @@ export default class ProfileSidebar extends Base {
         </div>
         <hr />
         <div className={css('info')}>
-          {/* <StatItemList stats={statItemList} style={{marginBottom: 40}} /> */}
           { /* TODO - these links break errrrythang profile.homeUrl ? <p><a href={profile.homeUrl}>{profile.homeUrl}</a></p> : undefined */ }
           { profile.homeUrl ? <p>{profile.homeUrl}</p> : undefined }
           { /* TODO - profile.twitter ? <p><a href={`https://twitter.com/${profile.twitter}`}>{profile.twitter}</a></p> : undefined */ }
