+++ lib/components/stylesheet/Stylesheet.js
@@ -12,7 +12,6 @@ import StatsLine from '../StatsLine'
 import TabPanel from '../TabPanel'
 import Search from '../Search'
 import StatItem from '../item/StatItem'
-import StatItemList from '../StatItemList'
 import Button from '../Button'
 import List from '../List'
 import Dropdown from '../Dropdown'
@@ -46,7 +45,7 @@ export default class Stylesheet extends Base {
 
   template (css) {
     const palette = defaultPalette
-    const { stats, statItem, statItemList } = this.props
+    const { stats, statItem } = this.props
     return (
       <div id='stylesheet' className={css('stylesheet')}>
         <div className={css('container')}>
@@ -124,7 +123,6 @@ export default class Stylesheet extends Base {
               <StatsLine stats={stats} muted />
               <StatsLine stats={stats} extraSpace large />
               <List data={statItem} component={StatItem} />
-              <StatItemList stats={statItemList} />
             </div>
           </div>
         </div>
@@ -265,20 +263,6 @@ Stylesheet.defaultProps = {
     {
       icon: 'link',
       title: 'transferred',
-      stat: '2.4Gib'
-    }
-  ],
-  statItemList: [
-    {
-      title: 'dataset',
-      stat: '14'
-    },
-    {
-      title: 'time online',
-      stat: '3:32:01'
-    },
-    {
-      title: 'repo size',
       stat: '2.4Gib'
     }
   ]
