+++ lib/routes.js
@@ -5,12 +5,10 @@ import { Switch, Route } from 'react-router-dom';
 import AppContainer from './containers/App'
 import DatasetsContainer from './containers/Datasets'
 import DatasetContainer from './containers/Dataset'
-import AddDatasetContainer from './containers/AddDataset'
 import ProfilesContainer from './containers/Profiles'
 import ProfileContainer from './containers/Profile'
 import Stylesheet from './components/stylesheet/Stylesheet'
-// import SearchResultsContainer from './containers/SearchResults'
-import MetadataEditorContainer from './containers/MetadataEditor'
+
 import EditorContainer from './containers/Editor'
 
 export default () => (
@@ -21,7 +19,6 @@ export default () => (
       <Route path="/edit"  component={EditorContainer} />
 
       <Route path="/datasets" component={DatasetsContainer} />
-      <Route path="/dataset/add" component={AddDatasetContainer} />
 
       <Route path='/profile' component={ProfileContainer} />
       <Route exact path='/profiles' component={ProfilesContainer} />
@@ -29,8 +26,6 @@ export default () => (
 
       <Route exact path="/stylesheet" component={Stylesheet} />
       <Route exact path="/search" component={DatasetsContainer} />
-      
-      <Route path='/:peername/:name/at/:network/:hash/meta/edit' component={MetadataEditorContainer} />
       
       <Route path='/:peername/:name/at/:network/:hash' component={DatasetContainer} />
       <Route path='/:peername/:name' component={DatasetContainer} />
